+++ i18n/qqq.json
@@ -19,7 +19,7 @@
 	"language-change": "Title of the language settings page",
 	"language-setting": "Title of the popup message in the language settings page",
 	"language-setting-message": "Message shown in the popup message in the language settings page",
-	"ok": "Label of the word \"ok\"",
+	"ok": "Label of the word \"ok\" shown in the popup message in the language settings page",
 	"softkey-read": "Text of the read option in the softkey. Used to read the selected article in the article preview.\n{{Identical|Read}}",
 	"settings-language": "Label of the language item list. Shown in the Settings menu, used in language setting.",
 	"settings-textsize": "Label of the Text size item list. Shown in the Settings menu, used in text size setting.",
