+++ src/hooks/index.js
@@ -7,4 +7,5 @@ export * from './useNavigation'
 export * from './usePopup'
 export * from './useScroll'
 export * from './useSearch'
+export * from './useSearchLanguage'
 export * from './useSoftkey'
