+++ style/language.less
@@ -1,5 +1,4 @@
 .language {
-
 	background-color: #3366cc;
 	text-align: center;
